% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bees}
\alias{bees}
\title{bee endpoints}
\format{
A data frame with 8531 rows and 31 variables:
\describe{
\item{test_id}{test id}
\item{chemical_name}{chemial name}
...
}
}
\source{
\url{https://cfpub.epa.gov/ecotox/}
}
\usage{
bees
}
\description{
A dataset containing the bee data from EPA ecotox database
diamonds.
}
\keyword{datasets}
