% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data.B}
\alias{Data.B}
\title{bee endpoints}
\format{
A data frame with 21909 rows and 36 variables:
\describe{
\item{test_id}{test id}
\item{chemical_name}{chemial name}
...
}
}
\source{
data/Copy of Brian-May2009.csv
}
\usage{
Data.B
}
\description{
A dataset containing the USEPA internal not for publication
}
\keyword{datasets}
